requirements.txt
Django==3.2.19
djangorestframework==3.14.0
python-dotenv==1.0.0
phonenumbers==8.13.11
django-phonenumber-field==7.1.0
Pillow==9.5.0
django-filter==23.2
psycopg2-binary==2.9.6
celery[redis]==5.3.0
flower==1.2.0
celery-singleton==0.3.1
django-redis==5.2.0
social-auth-app-django==5.2.0
